reg_t/sreg_t defination recover; cap_reg_t aux structure added

+++ fesvr/htif.cc
@@ -234,7 +234,7 @@ int htif_t::run()
     device_list.tick();
 
     if (!fromhost_queue.empty() && !mem.read_uint64(fromhost_addr)) {
-      mem.write_uint64(fromhost_addr, to_target(static_cast<uint64_t>(fromhost_queue.front())));
+      mem.write_uint64(fromhost_addr, to_target(fromhost_queue.front()));
       fromhost_queue.pop();
     }
   }
